refactor : 검증 클래스 분리 검증 부분을 StoreValidator클래스에 구현했습니다.
